+++ packages/core/src/inTimeout.test.ts
@@ -1,9 +1,15 @@
-import { nextTick } from "@lazy-promise/core";
+import { inTimeout } from "@lazy-promise/core";
 import { afterEach, beforeEach, expect, test, vi } from "vitest";
 
 const logContents: unknown[] = [];
+let logTime: number;
 
 const log = (...args: unknown[]) => {
+  const currentTime = Date.now();
+  if (currentTime !== logTime) {
+    logContents.push(`${currentTime - logTime} ms passed`);
+    logTime = currentTime;
+  }
   logContents.push(args);
 };
 
@@ -17,14 +23,11 @@ const readLog = () => {
 
 beforeEach(() => {
   vi.useFakeTimers();
-  vi.spyOn(process, "nextTick").mockImplementation((callback, ...args) => {
-    setTimeout(() => callback(...args));
-  });
+  logTime = Date.now();
 });
 
 afterEach(() => {
   vi.useRealTimers();
-  vi.restoreAllMocks();
   try {
     if (logContents.length) {
       throw new Error("Log expected to be empty at the end of each test.");
@@ -35,13 +38,14 @@ afterEach(() => {
 });
 
 test("resolve", () => {
-  nextTick().subscribe((value) => {
+  inTimeout(1000).subscribe((value) => {
     log("handleValue", value);
   });
   expect(readLog()).toMatchInlineSnapshot(`[]`);
   vi.runAllTimers();
   expect(readLog()).toMatchInlineSnapshot(`
     [
+      "1000 ms passed",
       [
         "handleValue",
         undefined,
@@ -51,7 +55,7 @@ test("resolve", () => {
 });
 
 test("cancel", () => {
-  nextTick().subscribe(() => {
+  inTimeout().subscribe(() => {
     log("handleValue");
   })();
   vi.runAllTimers();