+++ packages/core/src/index.ts
@@ -1,13 +1,16 @@
 export { all } from "./all";
-export { animationFrame } from "./animationFrame";
 export { any } from "./any";
 export { catchFailure } from "./catchFailure";
 export { catchRejection } from "./catchRejection";
 export { finalize } from "./finalize";
 export { fromEager } from "./fromEager";
 export { fromGenerator } from "./fromGenerator";
-export { idleCallback } from "./idleCallback";
-export { immediate } from "./immediate";
+export { inAnimationFrame } from "./inAnimationFrame";
+export { inIdleCallback } from "./inIdleCallback";
+export { inImmediate } from "./inImmediate";
+export { inMicrotask } from "./inMicrotask";
+export { inNextTick } from "./inNextTick";
+export { inTimeout } from "./inTimeout";
 export {
   box,
   failed,
@@ -19,8 +22,5 @@ export {
 export type { LazyPromiseError, LazyPromiseValue } from "./lazyPromise";
 export { log } from "./log";
 export { map } from "./map";
-export { microtask } from "./microtask";
-export { nextTick } from "./nextTick";
 export { race } from "./race";
-export { timeout } from "./timeout";
 export { toEager } from "./toEager";